Suppression developer fonctionnelle

+++ app/routes/developers.js
@@ -4,22 +4,25 @@ export default Route.extend({
 	model() {
 		return {
 			developers: this.store.findAll('developer'),
-			selectedDeveloper: undefined
-		}
+            selectedDeveloper: undefined
+        }
 	},
 	actions: {
-		removeDeveloperItem(id) {
-			this.store.findRecord('developer', id).then(function(developer) {
-				developer.deleteRecord();
-				developer.get('isDeleted');
-				developer.save();
-			});
-		},
 		updateDeveloper(data) {
 			this.store.findRecord('developer', data.id).then(function(developer) {
-				developer.set('name',data.developerName);
+				developer.set('identity',data.identity);
 				developer.save();
 			});
-		}
+        },
+        showDeleteModal(data) {
+            let tmp = confirm("Confirmer la suppression de "+data.identity+" ?");
+            if (tmp) {
+                this.store.findRecord('developer', data.id).then(function(developer) {
+                    developer.deleteRecord();
+                    developer.get('isDeleted');
+                    developer.save();
+                });
+            }
+        }
 	}
 });
